versao completa

+++ lib/tela_inicial.dart
@@ -13,8 +13,7 @@ class TelaInicialPersonalizada extends StatefulWidget {
   const TelaInicialPersonalizada({super.key});
 
   @override
-  _TelaInicialPersonalizadaState createState() =>
-      _TelaInicialPersonalizadaState();
+  _TelaInicialPersonalizadaState createState() => _TelaInicialPersonalizadaState();
 }
 
 class _TelaInicialPersonalizadaState extends State<TelaInicialPersonalizada> {
@@ -259,8 +258,8 @@ class _TelaInicialPersonalizadaState extends State<TelaInicialPersonalizada> {
               child: Center(
                 child: GridView.count(
                   crossAxisCount: 3,
-                  mainAxisSpacing: 8,  // Diminui o espaço entre as linhas
-                  crossAxisSpacing: 8,  // Diminui o espaço entre as colunas
+                  mainAxisSpacing: 16,
+                  crossAxisSpacing: 16,
                   children: [
                     _buildMenuItem(
                       icon: Icons.people,
@@ -303,7 +302,8 @@ class _TelaInicialPersonalizadaState extends State<TelaInicialPersonalizada> {
                         Navigator.push(
                           context,
                           MaterialPageRoute(
-                              builder: (context) => const TelaCalendario()),
+                              builder: (context) =>
+                                  const TelaCalendario()),
                         );
                       },
                     ),
@@ -330,15 +330,15 @@ class _TelaInicialPersonalizadaState extends State<TelaInicialPersonalizada> {
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
           CircleAvatar(
-            radius: 25, // Tamanho do ícone foi reduzido
+            radius: 30,
             backgroundColor: primaryColor, 
-            child: Icon(icon, size: 28, color: Colors.white), // Ícone menor
+            child: Icon(icon, size: 30, color: Colors.white),
           ),
-          const SizedBox(height: 4),  // Reduzindo o espaço entre ícone e texto
+          const SizedBox(height: 8),
           Text(
             label,
             textAlign: TextAlign.center,
-            style: const TextStyle(fontSize: 12),  // Tamanho da fonte reduzido
+            style: const TextStyle(fontSize: 14),
           ),
         ],
       ),
@@ -441,30 +441,77 @@ class _TelaListaEventosState extends State<TelaListaEventos> {
                         ),
                         elevation: 4,
                         child: ListTile(
+                          contentPadding: const EdgeInsets.symmetric(
+                              vertical: 10, horizontal: 16),
                           title: Text(
-                            event['event']!,
+                            event['event'] ?? 'Evento sem nome',
                             style: const TextStyle(
-                                fontWeight: FontWeight.bold),
+                              fontWeight: FontWeight.bold,
+                              fontSize: 18,
+                            ),
                           ),
                           subtitle: Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
-                              Text(
-                                'Hora: ${event['time']}',
-                                style: const TextStyle(fontSize: 14),
-                              ),
+                              Text('Hora: ${event['time']}',
+                                  style: const TextStyle(fontSize: 14)),
                               const SizedBox(height: 4),
                               Text(
                                 'Descrição: ${event['description']}',
-                                style: const TextStyle(fontSize: 14),
+                                style: TextStyle(
+                                    color: Colors.grey[700], fontSize: 14),
                               ),
                             ],
                           ),
                           trailing: IconButton(
-                            icon: const Icon(Icons.delete, color: Colors.red),
-                            onPressed: () {
-                              widget.onDeleteEvent(event);
-                              _deleteEventFromFirebase(event);
+                            icon: const Icon(
+                              Icons.delete,
+                              color: Colors.red,
+                            ),
+                            onPressed: () async {
+                              bool? confirmDelete = await showDialog<bool>(
+                                context: context,
+                                builder: (BuildContext context) {
+                                  return AlertDialog(
+                                    title: const Text('Excluir Evento'),
+                                    content: const Text(
+                                        'Você tem certeza de que deseja excluir este evento?'),
+                                    actions: <Widget>[
+                                      TextButton(
+                                        onPressed: () {
+                                          Navigator.of(context).pop(false);
+                                        },
+                                        child: const Text('Cancelar'),
+                                      ),
+                                      TextButton(
+                                        onPressed: () {
+                                          Navigator.of(context).pop(true);
+                                        },
+                                        child: const Text('Excluir',
+                                            style:
+                                                TextStyle(color: Colors.red)),
+                                      ),
+                                    ],
+                                  );
+                                },
+                              );
+                              if (confirmDelete == true) {
+                                await _deleteEventFromFirebase(event);
+                                widget.onDeleteEvent(event); 
+                                setState(() {
+                                  widget.events[eventDate]?.remove(event);
+                                  if (widget.events[eventDate]?.isEmpty ?? false) {
+                                    widget.events.remove(eventDate);
+                                  }
+                                });
+
+                                ScaffoldMessenger.of(context).showSnackBar(
+                                  const SnackBar(
+                                    content: Text('Evento excluído com sucesso!'),
+                                    backgroundColor: Colors.green,
+                                  ),
+                                );
+                              }
                             },
                           ),
                         ),
